Suggested docker.txt changes
I suggest breaking the previous "docker ps" command into "docker ps --latest" and "docker ps --all". --latest will show the most recently created containers. the --all command unlike the ps command by itself will show both running and stopped docker containers.

Next I add two for loops. One is for running through only active containers and the other is for running through all containers stopped and started. The docker top command will only work on running containers, so it requires the first for loop. The other sub-commands logs and inspect are valuable for all containers running or not.

I had to use " instead of ' for the log_cmd commands within the for loops or it did not interpret the $i for loop variable.

Lastly I suggest listing the contents of certain directories to provide additional information on containers as well as to verify the files aren't missing. The directories I suggest are /var/lib/docker/containers and /sys/fs/cgroup/*/system.slice/docker*. It's possible that /var/lib/docker may be useful all together, but often times it is a very large directory resulting in a much larger docker.txt file than I'm not sure would be desired. I find that generally I just need the containers subdirectory of the /var/lib/docker directory. The other directory is to get more direct information about how it is relating to the cgroups. I found that docker information was always in a system.slice folder of various subdirectories of /sys/fs/cgroup, and that's the reason for the wild-cards.