Suggested docker.txt changes
Per Jason's suggestions:

I'm unable to combine --all and --latest, however the creation date of each container is shown within the --all flag so someone could still deduce the latest container without having that command. 

I've combined the for loops, there will be error output from the top sub-command of ps on inactive containers, but the error is informational.

Lastly, I looked over the past standard options for the ls command and updated as requested.