suportconfig: Add tumbleweed check along SLE/Leap 15
Change all places where it is currently checking for version 15* to also
check for version 999 (tumbleweed). Without this patch, the following
output is printed:

...
Crash Info...                                Done
NTP...                                         PROC...                                      Done

ntp_info only gather logs NTP data when version 11, 12 and 15, but
Tumbleweed versions is the current YYYYMMDD of the last snapshot
installed. This patch fixes this by adding a check for version 999 in
everyplace that was checking just for version 15.

Signed-off-by: Marcos Paulo de Souza <marcos.souza.org@gmail.com>