supportconfig: Add collection of livepatch information
Collect information relating to livepatch available under the path
/sys/kernel/livepatch/. Each of the subdirectories following the path
is a kernel livepatching module and hosts the information about the
current patching state of the functions to be livepatched by the
module. The files layout exported by the livepatching subsystem is:
/sys/kernel/livepatch
/sys/kernel/livepatch/<patch>
/sys/kernel/livepatch/<patch>/enabled
/sys/kernel/livepatch/<patch>/transition
/sys/kernel/livepatch/<patch>/force
/sys/kernel/livepatch/<patch>/<object>
/sys/kernel/livepatch/<patch>/<object>/<function,sympos>

This patch collects values from the files, but last two are empty
directories whos paths are recorded instead. Also, introduces/enables
the configuration OPTION_KERNEL_LIVEPATCH in supportconfig.rc file.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.ibm.com>