memory_info: Show VMware memory balloon infomation
Customers observed a lot of OOM problems which are caused by
VMware memory balloon dirver which can reserve a lot of physical
memory in its internal memory pool. However, these reserved physical
memory are not tracked in /proc/meminfo. So /sys/kernel/debug/vmmemctl
would be used as a complement of /proc/meminfo for debugging this kind
of memory issues.

Signed-off-by: Firo Yang <firogm@gmail.com>