supportconfig: sssd_info avoid running lsof against all active processes
lsof can take a very long time and generate a large amount of output.
We should avoid running it against all active processes (the default
when options aren't specified) unless absolutely necessary.

We only run it when we can't find the sssd pid, but it's very possible
that sssd would be installed, but not running, in which case there
would not be any benefit to running it.

Simplify the logic to search the process list by pid, if we have one,
or by name if we don't, so we only run 'lsof -p <pid>' when against potential
sssd pids.

This avoids running a potentially expensive operation unnecessarily.