supportconfig: drm: cleanup output
This is a clean up the output of the drm info collection.

Check whether modetest and modeprint are present before use
to avoid errors if they're not installed

intel-gpu-tools consolidated intel_reg_dumper into intel_reg,
so check which one we have before use to avoid errors

Also, when wildcards are used in a list in a bash 'for' loop, it
will output the list string if nothing matches, so to avoid errors,
check for the existance of the file/directory before use.