Adapt for new versions of kdump
kdump 1.9+ does not have kdumptool dump_config and has a different
format of README.txt (missing space).

Modify supportconfig to work with the above changes.

Modify analyzevmcore to read KDUMP_SAVEDIR the same way supportconfig does.
this also makes it work with the new kdump versions and also will use the
default value if KDUMP_SAVEDIR is not specified in te config file.